Add files via upload
Hi, I'm Karan from batch MDTE07. Here I am going to show my first project, which is unfortunately incomplete because the time is not sufficient for me to complete this project within a reasonable time. So far, this programme has run successfully, but I can't see the results because the streamlit is taking half a day to load. So hereby, I'm submitting the project, and I'm pretty sure that it won't happen on my next project.    


                                                                                 Thankyou